# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API для онлайн-магазина кондитерских изделий",
        "title": "Bakery App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/",
    "paths": {
...
        "/dict/{dict}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает все не удалённые элементы справочника (isDeleted = false)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dictionary"
                ],
                "summary": "Получить элементы справочника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Тип справочника (например: position, accessGroup, subcategoryProduct, statuses, paymentType)",
                        "name": "dict",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DictionaryItem"
                            }
                        }
                    },
                    "400": {
                        "description": "неверный тип справочника",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавляет новый элемент в указанный справочник",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dictionary"
                ],
                "summary": "Создать элемент справочника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Тип справочника",
                        "name": "dict",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Элемент справочника",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DictionaryItem"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.DictionaryItem"
                        }
                    },
                    "400": {
                        "description": "неверный тип справочника",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/dict/{dict}/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Помечает элемент как удалённый (isDeleted = true)",
                "tags": [
                    "Dictionary"
                ],
                "summary": "Удалить элемент справочника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Тип справочника",
                        "name": "dict",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID элемента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "элемент справочника удалён",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "неверный тип справочника",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.DictionaryItem": {
            "description": "Универсальный элемент справочника (используется для всех типов справочников)",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isdeleted": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}